Split retention modules for repository gates
