+++ crates/oasis7/src/bin/oasis7_chain_runtime/execution_bridge/tests/mod.rs
@@ -25,6 +25,7 @@ use sha2::{Digest, Sha256};
 mod determinism;
 mod driver;
 mod driver_checkpoint;
+mod driver_checkpoint_profile;
 mod driver_observability;
 mod driver_storage_profile;
 mod replay;
